Extract importFaker in endPointSpecFactory
Make separate entries for client and server, so that the faker-lib can
be imported dynamically on the client, and always required on the server
