+++ src/coreModules/api/endpointSpecFactory/client.js
@@ -1,0 +1,5 @@
+const createBuildEndpointSpec = require('./createBuildEndpointSpec')
+
+const importFaker = () => import('json-schema-faker')
+
+module.exports = createBuildEndpointSpec({ importFaker })
