+++ src/coreModules/api/endpointSpecFactory/server.js
@@ -1,0 +1,6 @@
+const createBuildEndpointSpec = require('./createBuildEndpointSpec')
+const faker = require('json-schema-faker')
+
+const importFaker = () => Promise.resolve(faker)
+
+module.exports = createBuildEndpointSpec({ importFaker })
