+++ src/utilities/jsonSchema/createMockDataFromSchema.js
@@ -1,6 +1,8 @@
-const importFaker = () => require('json-schema-faker') // eslint-disable-line global-require
-
-module.exports = function createMockDataFromSchema(schema, models) {
+module.exports = function createMockDataFromSchema({
+  importFaker,
+  models,
+  schema,
+}) {
   return importFaker().then(jsf => {
     const modelsToUse = Object.keys(models).map(moduleKey => {
       return models[moduleKey]
