+++ src/utilities/jsonSchema/createMockGenerator.js
@@ -1,7 +1,11 @@
 const models = require('dina-schema/build/models.json')
 const createMockDataFromSchema = require('./createMockDataFromSchema')
 
-module.exports = function createMockGenerator({ model, schema: customSchema }) {
+module.exports = function createMockGenerator({
+  importFaker,
+  model,
+  schema: customSchema,
+}) {
   if (model && !models[model]) {
     throw new Error(`Unknown model: ${model}`)
   }
@@ -11,8 +15,13 @@ module.exports = function createMockGenerator({ model, schema: customSchema }) {
       'If model not provided have to provide customSchema (key schema)'
     )
   }
+
+  if (!importFaker) {
+    throw new Error(`Must provide 'importFaker' method`)
+  }
+
   const schema = models[model] || customSchema
   return () => {
-    return createMockDataFromSchema(schema, models)
+    return createMockDataFromSchema({ importFaker, models, schema })
   }
 }
